% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R
\name{apa}
\alias{apa}
\title{APA printers}
\usage{
apa(test_result, pl = TRUE)
}
\arguments{
\item{test_result}{a result of a statistical test as returned by a function from the `rstatix` package. Currently supported are `t_test()`, `chisq_test()`.}

\item{pl}{if `TRUE` then polish format is used}
}
\value{
a string with the APA formatted test result
}
\description{
APA printers
}
