% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_kable.R
\name{cat_kable}
\alias{cat_kable}
\title{cat names in a list as kable calls}
\usage{
cat_kable(list)
}
\arguments{
\item{list}{a named list}
}
\description{
cat names in a list as kable calls
}
\examples{
cat_kable(list(a = mtcars, b = iris))
}
