% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_percent_plot.R
\name{col_percent_plot}
\alias{col_percent_plot}
\title{Column percent plot}
\usage{
col_percent_plot(
  freq_df,
  x,
  nudge_y = -0.03,
  xlab = NULL,
  ylab = NULL,
  caption = NULL,
  pl = TRUE,
  fill = "#2171B5"
)
}
\arguments{
\item{freq_df}{a df returned by \code{freq_table()}. Must contain a column \code{prop}}

\item{x}{column to use for x axis}

\item{nudge_y}{vertical nudge for labels}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{caption}{caption}

\item{pl}{logical, whether to use Polish formatting}

\item{fill}{color for the plot}
}
\value{
ggplot object, a col plot showing a percentage in a group
}
\description{
Column percent plot
}
