% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{Frequency table}
\usage{
freq_table(...)
}
\arguments{
\item{...}{parameters passed to \code{rstatix::freq_table()}}
}
\value{
A tibble with frequency and proportion
}
\description{
A wrapper around \code{rstatix::freq_table()} with unrounded proportions. Useful for ggplot2.
}
