% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_plot.R
\name{lm_plot}
\alias{lm_plot}
\title{Plot for linear regression}
\usage{
lm_plot(df, x, y, xlab, ylab)
}
\arguments{
\item{df}{a data frame}

\item{x}{variable for x axis}

\item{y}{variable for y axis}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}
}
\description{
Plot for linear regression
}
