% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_upper_cor_mat.R
\name{lower_upper_cor_mat}
\alias{lower_upper_cor_mat}
\title{Łączenie macierzy korelacji}
\usage{
lower_upper_cor_mat(lower, upper, signif = TRUE)
}
\arguments{
\item{lower}{cor_mat. Macierz dolna.}

\item{upper}{cor_mat. Macierz górna.}

\item{signif}{logical. Jeśli TRUE, dodaje gwiazdki na oznaczenie istotności.}
}
\value{
Zwraca obiekt typu \code{\link[tibble]{tibble}}.
}
\description{
\code{lower_upper_cor_mat} łączy dwie macierze korelacji tak, że jedna połowa
jest pod przekątną, a druga nad przekątną. Użyteczne w raportowaniu. W
przeciwieństwie do klasycznego \code{\link[psych]{lowerUpper}}, ta
obsługuje obiekt typu \code{\link[rstatix]{cor_mat}}.
}
\examples{
library(dplyr)
library(rstatix)

r_mat_setosa <- iris \%>\%
  filter(Species == "setosa") \%>\%
  select(where(is.numeric)) \%>\%
  cor_mat()

r_mat_virginica <- iris \%>\%
  filter(Species == "virginica") \%>\%
  select(where(is.numeric)) \%>\%
  cor_mat()

lower_upper_cor_mat(r_mat_setosa, r_mat_virginica)
}
\seealso{
\code{\link[psych]{lowerUpper}}, \code{\link[rstatix]{cor_mat}}
}
