% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_plot.R
\name{M_plot}
\alias{M_plot}
\title{Means plot}
\usage{
M_plot(
  df,
  group,
  col,
  x_labels,
  xlab = NULL,
  ylab = NULL,
  nudge_y = -1,
  fill = "#1E88E5"
)
}
\arguments{
\item{df}{a data frame}

\item{group}{group variable to summarise by}

\item{col}{column to summarise}

\item{x_labels}{vector of labels for x axis}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{nudge_y}{vertical nudge for labels}
}
\value{
ggplot object, a bar plot showing means with error bars (95\% CI)
}
\description{
Means plot
}
