% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSD.R
\name{msd}
\alias{msd}
\title{Report M and SD}
\usage{
msd(df, column, group, pl = TRUE)
}
\arguments{
\item{df}{data frame}

\item{column}{column of which to calculate M and SD}

\item{group}{optional grouping variable}
}
\value{
Ciąg znaków zgodny z przykładem:
"$M = 5,01$; $SD = 0,35$"
}
\description{
Report M and SD
}
\examples{
msd(iris, Sepal.Length)
msd(iris, Sepal.Length, Species == "setosa")
}
