% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_plot.R
\name{prop_plot}
\alias{prop_plot}
\title{Proportions plot}
\usage{
prop_plot(
  freq_df,
  legend_title,
  x = 1,
  order = FALSE,
  xlab = NULL,
  ylab = NULL,
  fill_palette = "Blues",
  ...
)
}
\arguments{
\item{freq_df}{a df returned by \code{freq_table()}}

\item{legend_title}{title for the legend}

\item{x}{column to use for x axis}

\item{order}{logical, whether to order the groups by proportion}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{fill_palette}{RColorBrewer color palette name for the plot}
}
\value{
ggplot object, a plot of proportions summed to 100\%, horizontal
}
\description{
Percent stacked bar plot with labels. Should be exported with \code{width = 12} and \code{height = 4}.
}
