% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_cor.R
\name{raport_cor}
\alias{raport_cor}
\title{Raport automatyczny z korelacji}
\usage{
raport_cor(wynik_testu, miedzy, method = "pearson", alpha = 0.05)
}
\arguments{
\item{wynik_testu}{\code{\link[stats:cor.test]{stats::cor.test()}}.}

\item{miedzy}{character. Uzupełnia zdanie „Przeprowadzono analizę korelacji między XXX.”}

\item{method}{character. Rodzaj korelacji – "pearson" albo "spearman"}

\item{alpha}{numeric. Punkt odcięcia dla istotności statystycznej.}
}
\value{
Przeprowadzono analizę korelacji Pearsona między masą a ceną
brylantu. Test wykazał istnienie istotnej, bardzo wysokiej, dodatniej
korelacji między zmiennymi ($r = 0,922$; $p < 0,001$).
}
\description{
Funkcja tworząca automatyczne fragmenty raportu z analizy korelacji.
}
\examples{
data(diamonds, package = "ggplot2")
r <- cor.test(diamonds$carat, diamonds$price)
raport_cor(r, "masą a ceną brylantu")

}
