% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_KW.R
\name{raport_KW}
\alias{raport_KW}
\title{Raport automatyczny z testu Kruskala-Wallisa}
\usage{
raport_KW(kruskaltest, roznic)
}
\arguments{
\item{roznic}{Uzupełnienie zdania "Celem sprawdzenia różnic "}

\item{kruskal.test}{obiekt testu Kruskala-Wallisa \code{\link[stats]{kruskal.test}}}
}
\value{
Celem sprawdzenia różnic w zakresie lęku u osób o różnym stopniu obciążenia
pracą przeprowadzono jednoczynnikową analizę wariancji. Wykazała ona, że istnieją istotne
różnice (F(2,597) = 5,22; p = 0,006).
}
\description{
Funkcja tworząca automatyczne fragmenty raportu z testu Kruskala-Wallisa
}
\examples{
iris_KW <- kruskal.test(Sepal.Length ~ Species, iris)
raport_KW(iris_KW, "w zakresie długości płatków u różnych gatunków irysów")
}
