% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_lm.R
\name{raport_lm}
\alias{raport_lm}
\title{Raport automatyczny z regresji}
\usage{
raport_lm(lm, miedzy, alpha = 0.05)
}
\arguments{
\item{lm}{\code{\link[stats]{lm}}.}

\item{miedzy}{character. Uzupełnia zdanie „Celem sprawdzenia zależności między XXX.”}

\item{alpha}{numeric. Punkt odcięcia dla istotności statystycznej.}
}
\value{
Celem sprawdzenia zależności między masą brylantu a jego ceną, przeprowadzono
analizę regresji. Model okazał się istotny statystycznie (F(1;53938) = 30410; p < 0,001)
i wyjaśnia 85% wariancji.
}
\description{
Funkcja tworząca automatyczne fragmenty raportu z analizy regresji.
}
\examples{
data(diamonds, package = "ggplot2")
llm <- lm(price ~ carat, diamonds)
raport_lm(llm, "masą brylantu a jego ceną")

}
