% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_t.R
\name{raport_t}
\alias{raport_t}
\title{Raport automatyczny z testu t}
\usage{
raport_t(wynik_testu, miedzy, alpha = 0.05)
}
\arguments{
\item{wynik_testu}{\code{\link[stats:t.test]{stats::t.test()}}.}

\item{miedzy}{character. Uzupełnia zdanie „Celem sprawdzenia istotności różnicy między XXX.”}

\item{alpha}{numeric. Punkt odcięcia dla istotności statystycznej.}
}
\value{
Celem sprawdzenia istotności różnicy między długością płatków u
różnych gatunków irysów wykonano test t-Studenta dla prób niezależnych.
Test wykazał, że różnica w średnich (ΔM = 2,798) między grupami jest
istotna statystycznie (t(62,1) = 39,5; p < 0,001).
}
\description{
Funkcja tworząca automatyczne fragmenty raportu z testu t-Studenta.
}
\examples{
library(dplyr)
t <- iris \%>\%
  filter(Species \%in\% c("setosa", "versicolor")) \%>\%
  t.test(Petal.Length ~ Species, .)
raport_t(t, "długością płatków u różnych gatunków irysów")

}
