% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_U.R
\name{raport_U}
\alias{raport_U}
\title{Raport automatyczny z testu U}
\usage{
raport_U(wynik_testu, miedzy, alpha = 0.05)
}
\arguments{
\item{wynik_testu}{\code{\link[coin:wilcox_test]{coin::wilcox_test}}.}

\item{miedzy}{character. Uzupełnia zdanie „Przeprowadzono analizę korelacji między XXX.”}

\item{alpha}{numeric. Punkt odcięcia dla istotności statystycznej.}
}
\value{
Celem sprawdzenia istotności różnicy między długością płatków u
  różnych gatunków irysów wykonano test U Manna-Whitneya. Test wykazał, że
  różnica jest istotna statystycznie (Z = -7,468; p < 0,001).
}
\description{
Funkcja tworząca automatyczne fragmenty raportu z testu U Manna-Whitneya.
}
\examples{
library(dplyr)
U <- iris \%>\% filter(Species \%in\% c("setosa", "versicolor")) \%>\% coin::wilcox_test(Sepal.Length ~ Species, .)
raport_U(U, "długością płatków u różnych gatunków irysów")
}
