% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statystyki_opisowe.R
\name{statystyki_opisowe}
\alias{statystyki_opisowe}
\alias{statystyki_opisowe_by}
\alias{opisz}
\alias{opisz_by}
\title{Statystyki opisowe}
\usage{
statystyki_opisowe(
  df,
  type = c("kategorialna", "porzadkowa", "ilosciowa"),
  ...
)

statystyki_opisowe_by(
  df,
  type = c("kategorialna", "porzadkowa", "ilosciowa"),
  group,
  ...
)

opisz(df, kategorialne, porzadkowe, ilosciowe)

opisz_by(df, group, kategorialne, porzadkowe, ilosciowe)
}
\arguments{
\item{df}{data.frame lub tibble. Zbiór danych.}

\item{type}{character. Skala pomiarowa. Może być \code{"kategorialna"}, \code{"porzadkowa"} albo \code{"ilosciowa"}.}

\item{...}{tidyselect. Kolumny do opisania. Jeśli nie sprecyzowano, opisywane są wszystkie dostępne kolumny. Można użyć czasowników tidyselect jak \code{\link[dplyr]{starts_with}}.}

\item{group}{kolumny ze zmienną grupującą. Podane jako nazwa, ciąg znaków lub wektor tychże.}

\item{kategorialne}{Wektor nazw zmiennych kategorialnych w ramce danych.}

\item{porzadkowe}{Wektor nazw zmiennych porządkowych w ramce danych.}

\item{ilosciowe}{Wektor nazw zmiennych ilościowych w ramce danych.}
}
\value{
Zwracana jest nazwana lista, której zawartość zależy od skali pomiarowej.
\describe{
  \item{kategorialna}{\describe{
    \item{licznosc}{tabele liczności}
  }}
  \item{porządkowa}{\describe{
    \item{licznosc}{tabele liczności}
    \item{opisowe}{\code{var} – nazwa zmiennej; \code{N} – liczba obserwacji; \code{Me} – mediana; \code{R} – rozstęp; \code{RQ} – rozstęp ćwiartkowy; \code{VQ} – współczynnik zmienności wzg. mediany; \code{VQ_interp} – interpretacja \code{VQ}; \code{`NA`} – liczba braków danych}
  }}
  \item{ilościowa}{\describe{
    \item{opisowe}{\code{var} – nazwa zmiennej; \code{N} – liczba obserwacji; \code{M} – średnia; \code{SD} – odchylenie standardowe; \code{A} – skośność; \code{K} – kurtoza; \code{`NA`} – liczba braków danych}
    \item{test_shapiro_wilka}{test normalności Shapiro-Wilka z dodaną istotnością}
  }}
}

Nazwana lista statystyk opisowych dla zmiennych każdego typu.

Nazwana lista statystyk opisowych dla zmiennych każdego typu.
}
\description{
Zestaw podstawowych statystyk dla danego typu skali pomiarowej, zgodny z
wymaganiami prowadzących. \code{statystyki_opisowe} służy do opisu całych
sprecyzowanych kolumn. \code{statystyki_opisowe_by} robi to samo, ale pozwala
na dodanie zmiennych grupujących.
}
\section{Functions}{
\itemize{
\item \code{statystyki_opisowe_by}: Statystyki opisowe z grupowaniem.

\item \code{opisz}: Opisuje całą ramkę danych, użytkownik wskazuje zmienne danego typu.

\item \code{opisz_by}: Opisuje całą ramkę danych ze zmienną grupującą, użytkownik wskazuje zmienne danego typu.
}}

\examples{
statystyki_opisowe(iris, "ilosciowa", Sepal.Width)
statystyki_opisowe(iris, "ilosciowa", tidyselect:::where(is.numeric)) # z czasownikiem tidyselect

data(diamonds, package = "ggplot2")
# wiele kolumn do opisania
statystyki_opisowe_by(diamonds, "porzadkowa", group = clarity, price, carat)
statystyki_opisowe_by(diamonds, "kategorialna", c(cut, clarity), color)

data(diamonds, package = "ggplot2")
opisz(head(diamonds), color, table, c(carat, price))

data(diamonds, package = "ggplot2")
opisz_by(head(diamonds, 100), cut, color, table, c(carat, price))

}
