% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\title{Tidy wrapper}
\usage{
tidy(...)
}
\arguments{
\item{...}{arguments to pass to \code{broom::tidy()}}
}
\value{
a tibble with additional columns for significance
}
\description{
Tidy wrapper
}
