% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_char.R
\name{vector_to_char}
\alias{vector_to_char}
\alias{name_to_char}
\title{Zmiana nazw w ciągi znaków}
\usage{
vector_to_char(vctr_of_names)

name_to_char(name)
}
\arguments{
\item{vctr_of_names}{wektor nazw. Może zawierać także ciągi znaków.}

\item{name}{nazwa.}
}
\value{
Zwraca ciąg znaków lub nienazwany wektor ciągów znaków.
}
\description{
Funkcje zmieniające nazwy w ciągi znaków, potencjalnie przydatne w
programowaniu. Zdejmują z użytkownika konieczność umieszczania wejścia w
cudzysłowach.
}
\section{Functions}{
\itemize{
\item \code{name_to_char}: Pojedyncza nazwa
}}

\examples{
vector_to_char(c(plec, wiek))
vector_to_char(c(plec, "wiek")) # dopuszczalne, niezalecane
#' name_to_char(plec)
}
