% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jedrusiakr.R
\docType{package}
\name{jedrusiakr}
\alias{jedrusiakr}
\alias{jedrusiakr-package}
\title{jedrusiakr: Random Tools I Made and Use}
\description{
A set of some random tools I created for myself to use.
}
\author{
\strong{Maintainer}: Jakub Jędrusiak \email{jakub@s-el.pl}

}
