% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jedrusiakr.R
\docType{package}
\name{jedrusiakr}
\alias{jedrusiakr}
\alias{jedrusiakr-package}
\title{jedrusiakr: Utilities for Statistics in Psychology}
\description{
A set of some random tools I created for myself to use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jakub-jedrusiak/jedrusiakr}
  \item Report bugs at \url{https://github.com/jakub-jedrusiak/jedrusiakr/issues}
}

}
\author{
\strong{Maintainer}: Jakub Jędrusiak \email{jakub@s-el.pl}

}
