% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jedrusiakr.R
\docType{package}
\name{jedrusiakr}
\alias{jedrusiakr-package}
\alias{jedrusiakr}
\title{jedrusiakr: Utilities for Statistics in Psychology}
\description{
A set of tools for quick exploration and reporting. The main part consists of
functions used to calculate basic descriptive statistics and normality checks
(mainly `opisz()` and `opisz_by()`). The types of descriptives reported are
different for different types of measurement scales (nominal, ordinal and
continous). The package also includes a set of functions for autoreporting
(`raport_()`) that take a model and returns an example of analysis
description in Polish. For the time being, the package
and its documentation are written only in Polish.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jakub-jedrusiak/jedrusiakr}
  \item Report bugs at \url{https://github.com/jakub-jedrusiak/jedrusiakr/issues}
}

}
\author{
\strong{Maintainer}: Jakub Jędrusiak \email{jakub@s-el.pl}

}
