% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_plot.R
\name{lm_plot}
\alias{lm_plot}
\title{Plot for linear regression}
\usage{
lm_plot(
  df,
  x,
  y,
  ...,
  xlab = NULL,
  ylab = NULL,
  line_colour = "#1E88E5",
  point_colour = "#9ECAE1"
)
}
\arguments{
\item{df}{a data frame}

\item{x}{variable for x axis}

\item{y}{variable for y axis}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> additional arguments to pass to \code{aes()}}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}
}
\description{
Plot for linear regression
}
