% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSD.R
\name{MSD}
\alias{MSD}
\title{Report M and SD}
\usage{
MSD(df, column, group)
}
\arguments{
\item{var}{nazwa zmiennej do opisania, bez cudzysłowu}
}
\value{
Ciąg znaków zgodny z przykładem:
"$M = 5,01$; $SD = 0,35$"
}
\description{
Report M and SD
}
\examples{
MSD(iris, Sepal.Length)
MSD(iris, Sepal.Length, Species == "setosa")
}
