% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport_KW.R
\name{raport_KW}
\alias{raport_KW}
\title{Raport automatyczny z testu Kruskala-Wallisa}
\usage{
raport_KW(kruskaltest, roznic)
}
\arguments{
\item{kruskaltest}{obiekt testu Kruskala-Wallisa \code{\link[stats]{kruskal.test}}}

\item{roznic}{Uzupełnienie zdania "Celem sprawdzenia różnic "}
}
\value{
Celem sprawdzenia różnic w zakresie długości płatków u różnych gatunków
irysów przeprowadzono nieparametryczny test Kruskala-Wallisa. Wykazał on, że
istnieją istotne statystycznie różnice ($\\chi^2 (2) = 96,94$; $p < 0,001$).
}
\description{
Funkcja tworząca automatyczne fragmenty raportu z testu Kruskala-Wallisa
}
\examples{
iris_KW <- kruskal.test(Sepal.Length ~ Species, iris)
raport_KW(iris_KW, "w zakresie długości płatków u różnych gatunków irysów")
}
